+++ docs/user_guide/01_material-handling/temperature-controllers/temperature-controllers.rst
@@ -126,3 +126,4 @@ e.g.:
 
    ot-temperature-controller
    hamilton-heater-cooler
+   inheco
