+++ pylabrobot/heating_shaking/__init__.py
@@ -2,9 +2,7 @@
 
 from pylabrobot.heating_shaking.backend import HeaterShakerBackend
 from pylabrobot.heating_shaking.chatterbox import HeaterShakerChatterboxBackend
-from pylabrobot.heating_shaking.hamilton_backend import (
-  HamiltonHeaterShakerBackend,
-  HamiltonHeaterShakerBox,
-)
+from pylabrobot.heating_shaking.hamilton_backend import HamiltonHeaterShakerBackend, HamiltonHeaterShakerBox
 from pylabrobot.heating_shaking.heater_shaker import HeaterShaker
-from pylabrobot.heating_shaking.inheco_backend import InhecoThermoShakeBackend
+from pylabrobot.heating_shaking.inheco.thermoshake_backend import InhecoThermoshakeBackend
+from pylabrobot.heating_shaking.inheco.thermo_shake import inheco_thermoshake, inheco_thermoshake_ac, inheco_thermoshake_rm
