+++ pylabrobot/temperature_controlling/inheco/cpac_backend.py
@@ -1,0 +1,7 @@
+from pylabrobot.temperature_controlling.inheco.temperature_controller import (
+  InhecoTemperatureControllerBackend,
+)
+
+
+class InhecoCPACBackend(InhecoTemperatureControllerBackend):
+  pass
